Add files via upload
Temperature = 0 Celsius - value from paper in drive 

Altitude = 500 hPa - altitude is between the pressure at sea level and at the top of Everest. 5.5km above sea level, maybe an overestimate based on the service areas we were talking about but can help us justify our design

Wind gusts are added

the environment is a collection of mid-range hills we can use for testing the drone flying around obstacles. For the final simulation this can be replaced with a larger mountain.